bug(backend): prevent duplicate emails
After registration, the code tried to log a user in.  An unverified user
would not be able to login.  It would automatically send a new
verification email in that situation.  But, it would have just sent a
verification email as part of the registration process.

To address this, I stopped automatically sending a new verification
email during login.  And I added a new endpoint for sending a new
verification email.  In order for that to work, you need to pass it a
valid email address.  So, the error responses for unverified (and
unapproved) users now includes some minimal information about the user.

I also added a verify action to the user admin.

IssueID #27