feat(backend): update migration to use uuid as customer.pk
Updated the migration to use a UUIDField as the Customer primary key;
The original code used a blank db, but when trying to use this library
in an existing project it fell over trying to cast an integer as a uuid.
This updated migration, does it in several steps: creating the uuid,
then deleting the existing pk, then setting the new uuid as the pk.