feat(backend): resend invitation
Added a new model method to send an invitation to a CustomerUser.
Deleted the code that automatically did that as part of the
CustomerUserSerializer save method, and explicitly called `invite` as
part of the CustomerUser create view.  Also added another view just for
resending the invitation.  Added some tests.

IssueID #77