chore(backend): fixed a missed error in PR #105
The new registration_stage field was misspelt in the swagger
documentation.  More importantly, the registration_stage serializer
field in the RegisterSerializer didn't correctly validate against choices
(this was b/c the RegistrationSerializer was not a ModelSerializer based
on User where field attributes are set automatically based on the
corresponding model fields).