feat(backend): refactored onboarding
Added a view to onboard a user so it can be called by the frontend.
Removed default onboarding templates so an email is not automatically
sent.
Moved the onboarding field to the BasicUserSerializer (on the assumption
that a Customer Manager will want to know if their users have been
onboarded or not).