feat(backend): force accepted_terms to be false for new CustomerUsers
Gets around the problem of requiring accepted_terms to be True in order
for registration to work, but needing it to be False in order to require users created via an
invitation to a cutomer to explicitly accept_terms when logging in.  Did
this by setting it to True when validating the RegistrationSerializer
and then immediately setting it to False before saving the
RegistrationSerializer.