feat(backend): added support for messages
Added a message class to record emails sent to users in the db.
Provided the standard model/admin/serializer/view code.  Made the API
part of the users namespace: "/api/users/<user_id>/messages/" and did
not include messages in the normal user serialization; it must be
accesssed separately.  Also, ensured that most of the serializer fields
are read-only so we don't inadvertently change a message after-the-fact.

Also fixed the "example" app to use an appropriate DEFAULT_AUTO_FIELD
(so it doesn't generate spurious migrations).

Also some linting.