chore(backend): moved notification code from adapter to form/serializer
dj_rest_auth recently changed the way users are registered; They added
additional checks to the process.  This means that the adapter code that
creates a user doesn't actually always save that user anymore - saving
is done in the serializer after the additional checks.  Therefore, my
notification code was never running.  This commit moves the notification
code into the serializer to be called after the new checks have passed.
I added similar code in the form.